# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ChainLance API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ESCROW
...
  /conversations:
    get:
      operationId: listConversations
      tags: [messages]
      summary: List conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [messages]
      summary: Start a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get messages in conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # REVIEWS
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        role: { type: string }
        avatarUrl: { type: ["string", "null"] }
        walletAddress: { type: ["string", "null"] }
        isSuspended: { type: boolean }
        createdAt: { type: string }
        chainCoinBalance: { type: number }
        level: { type: ["integer", "null"] }
        trustScore: { type: ["number", "null"] }
...
    Freelancer:
      type: object
      required: [id, userId, bio, skills, category, hourlyRate, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        user:
          $ref: "#/components/schemas/User"
        bio: { type: string }
        skills:
          type: array
          items: { type: string }
        category: { type: string }
        hourlyRate: { type: number }
        status: { type: string }
        portfolioItems:
          type: array
          items: { type: string }
        completedOrders: { type: integer }
        avgRating: { type: ["number", "null"] }
        totalEarnings: { type: number }
        level: { type: ["integer", "null"] }
        createdAt: { type: string }
...
    Service:
      type: object
      required: [id, freelancerId, title, description, category, price, deliveryDays, status, createdAt]
      properties:
        id: { type: integer }
        freelancerId: { type: integer }
        freelancer:
          $ref: "#/components/schemas/Freelancer"
        title: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number }
        deliveryDays: { type: integer }
        status: { type: string }
        thumbnailUrl: { type: ["string", "null"] }
        previewImages:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        totalOrders: { type: integer }
        avgRating: { type: ["number", "null"] }
        isFeatured: { type: boolean }
        createdAt: { type: string }
...
    Order:
      type: object
      required: [id, serviceId, clientId, freelancerId, amount, status, createdAt]
      properties:
        id: { type: integer }
        serviceId: { type: integer }
        service:
          $ref: "#/components/schemas/Service"
        clientId: { type: integer }
        client:
          $ref: "#/components/schemas/User"
        freelancerId: { type: integer }
        freelancer:
          $ref: "#/components/schemas/Freelancer"
        amount: { type: number }
        status: { type: string }
        requirements: { type: ["string", "null"] }
        deliverables: { type: ["string", "null"] }
        dueDate: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        escrowId: { type: ["integer", "null"] }
        createdAt: { type: string }
    OrderInput:
      type: object
      required: [serviceId, requirements]
      properties:
        serviceId: { type: integer }
        requirements: { type: string }
    OrderUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
        deliverables: { type: string }
        rejectionReason: { type: string }
    OrderList:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ESCROW
...
    Conversation:
      type: object
      required: [id, participantIds, createdAt]
      properties:
        id: { type: integer }
        participantIds:
          type: array
          items: { type: integer }
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount: { type: integer }
        createdAt: { type: string }
    ConversationInput:
      type: object
      required: [participantId]
      properties:
        participantId: { type: integer }
        orderId: { type: ["integer", "null"] }
    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id: { type: integer }
        conversationId: { type: integer }
        senderId: { type: integer }
        sender:
          $ref: "#/components/schemas/User"
        content: { type: string }
        fileUrl: { type: ["string", "null"] }
        fileType: { type: ["string", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string }
    MessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
        fileUrl: { type: string }
        fileType: { type: string }

    # REVIEWS
...
    Notification:
      type: object
      required: [id, userId, type, title, message, isRead, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        isRead: { type: boolean }
        linkUrl: { type: ["string", "null"] }
        createdAt: { type: string }

    # ANALYTICS
...