# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChainLance API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Freelancer } from "./freelancer";

export interface Service {
  id: number;
  freelancerId: number;
  freelancer?: Freelancer;
  title: string;
  description: string;
  category: string;
  price: number;
  deliveryDays: number;
  status: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  previewImages?: string[];
  tags?: string[];
  totalOrders?: number;
  /** @nullable */
  avgRating?: number | null;
  isFeatured?: boolean;
  createdAt: string;
}